[Feature] newProduct back-end updated
