+++ .env.exemple
@@ -1,0 +1,3 @@
+PORT=
+MONGO_URI=
+JWT_SECRET=
