+++ src/index.js
@@ -4,7 +4,7 @@ import cors from 'cors';
 import chalk from 'chalk';
 import authRoutes from './routes/authRoutes.js'
 import cartRoutes from './routes/cartRoutes.js';
-import productRoutes from './routes/productRoutes.js'
+import productsRoutes from './routes/productsRoutes.js'
 
 dotenv.config();
 
@@ -16,9 +16,9 @@ app.get('/', (req, res) => {
     res.send('Hello World');
 });
 
-app.use(cartRoutes)
+app.use(cartRoutes);
 app.use(authRoutes); 
-app.use(productRoutes); 
+app.use(productsRoutes); 
 
 app.listen(process.env.PORT, () => {
     console.log(chalk.bold.red('Server running on port ' + process.env.PORT));
